:sparkles: Keep task name in the "Add Item" after screen rotation
Addition of a simple (but powerful) support for keeping the task name
inserted in the "Add Item" after a screen rotation or app kill by
Android system.

In order to achieve this, Koin was updated to support `stateViewModel`
and the `ViewModel` itself receives an `SavedStateHandle` which stores
the data. This objects is listened every time the screen is recreated
and set the value on the field if it is available.