:bug: Fix LiveData not update during screen rotation
In the previous code, the LiveData of each ViewModel was assigned during
construction. It means that the reference would be the same for the
older object after screen rotation and once it is attached to the
fragment lifecycle, it will stop listening to changes.

By updating it to a custom getter, every time the object is accessed it
will redirect to the latest instance of LiveData and receive all the
new changes.