:building_construction: Update Toolbar title update mechanism
The previous Toolbar title update mechanism relied in two different
strategies: The fragment label from the Navigation Graph or the value
sent via `LiveData` in the `SharedViewModel` which updated the
`MainActivity`. The problem is that now that Alkaa support landscape,
when the screen is rotate and the views are restored, the `LiveData`
may trigger an update in the title that may not be necessary.

In order to avoid this undesired behavior, the code is not updated to
always rely in the value send via the `LiveData`. I know that it
basically add more complexity for each fragment and new ones needs to
implement it, but this will be the way to go in order to avoid multiple
points of update as previously.